--- src/components/Header.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import logo from '../logo.svg';
-
-function Header() {
-  return (
-    <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />        
-    </header>
-  );
-}
-
-export default Header;
