+++ src/container/home/Home.js
@@ -3,7 +3,7 @@ import React from 'react';
 function Home() {
   return (
     <div className="mainWrapper d-flex flex-column justify-content-around">        
-        <div className="AppContent"><img src={require('../assets/images/testImage.jpg')} alt="Content Image"/></div>        
+        <div className="AppContent"><img src={require('../../assets/images/testImage.jpg')} alt="Content Image"/></div>        
     </div>
   );
 }
